tweaks for linuxkm targeting clang-built kernels:
linuxkm/: when logging PTR_ERR(), cast it to int, and use "%d" as the format.  Globally, `#define PTR_ERR(x) ((int)PTR_ERR(x))` in linuxkm_wc_port.h to fix clang warnings on kernel headers.

linuxkm/lkcapi_aes_glue.c: add casts in linuxkm_test_aesgcm() to mollify clang.

linuxkm/linuxkm_wc_port.h, linuxkm/module_hooks.c:
* add __clang__ compat code to allow including clang stdatomic.h while masking out kernel-incompatible __CLANG_STDINT_H.
* add clang-specific suppressions for kernel headers (-Wshorten-64-to-32, -Wframe-address).

linuxkm/lkcapi_sha_glue.c:
* in wc__get_random_bytes(), add bounds-checking for len.
* in wc_extract_crng_user(), fix type conflicts.

wolfssl/wolfcrypt/wc_port.h and wolfssl/wolfcrypt/types.h:
* move the old-FIPS compatibility mapping from INLINE to WC_INLINE from types.h to wc_port.h.
* activate stdatomic.h for clang kernel module builds.

linuxkm/Kbuild:
* add clang-specific flags.
* add gcc gate around gcc-specific flags.
* allow override value for MAX_STACK_FRAME_SIZE.

wolfcrypt/src/asn.c: add casts in GetFormattedTime_ex() to mollify clang build of linuxkm.

+++ linuxkm/lkcapi_glue.c
@@ -125,8 +125,8 @@ WC_MAYBE_UNUSED static int check_skcipher_driver_masking(struct crypto_skcipher 
         tfm = crypto_alloc_skcipher(alg_name, 0, 0);
     }
     if (IS_ERR(tfm)) {
-        pr_err("error: allocating skcipher algorithm %s failed: %ld\n",
-               alg_name, PTR_ERR(tfm));
+        pr_err("error: allocating skcipher algorithm %s failed: %d\n",
+               alg_name, (int)PTR_ERR(tfm));
         return -EINVAL;
     }
     actual_driver_name = crypto_tfm_alg_driver_name(crypto_skcipher_tfm(tfm));
@@ -158,8 +158,8 @@ WC_MAYBE_UNUSED static int check_aead_driver_masking(struct crypto_aead *tfm, co
         tfm = crypto_alloc_aead(alg_name, 0, 0);
     }
     if (IS_ERR(tfm)) {
-        pr_err("error: allocating AEAD algorithm %s failed: %ld\n",
-               alg_name, PTR_ERR(tfm));
+        pr_err("error: allocating AEAD algorithm %s failed: %d\n",
+               alg_name, (int)PTR_ERR(tfm));
         return -EINVAL;
     }
     actual_driver_name = crypto_tfm_alg_driver_name(crypto_aead_tfm(tfm));
@@ -191,8 +191,8 @@ WC_MAYBE_UNUSED static int check_shash_driver_masking(struct crypto_shash *tfm, 
         tfm = crypto_alloc_shash(alg_name, 0, 0);
     }
     if (IS_ERR(tfm)) {
-        pr_err("error: allocating shash algorithm %s failed: %ld\n",
-               alg_name, PTR_ERR(tfm));
+        pr_err("error: allocating shash algorithm %s failed: %d\n",
+               alg_name, (int)PTR_ERR(tfm));
         return -EINVAL;
     }
     actual_driver_name = crypto_tfm_alg_driver_name(crypto_shash_tfm(tfm));
